+++ web-admin/src/routes/[organization]/-/users/groups/+page.svelte
@@ -1,19 +1,14 @@
-<!-- <script lang="ts">
+<script lang="ts">
   import { page } from "$app/stores";
   import {
     createAdminServiceAddOrganizationMemberUsergroup,
-    createAdminServiceAddUsergroupMemberUser,
     createAdminServiceDeleteUsergroup,
     createAdminServiceGetCurrentUser,
     createAdminServiceListOrganizationMemberUsergroups,
     createAdminServiceListOrganizationMemberUsers,
     createAdminServiceRemoveOrganizationMemberUsergroup,
-    createAdminServiceRemoveUsergroupMemberUser,
-    createAdminServiceRenameUsergroup,
     createAdminServiceSetOrganizationMemberUsergroupRole,
     getAdminServiceListOrganizationMemberUsergroupsQueryKey,
-    getAdminServiceListOrganizationMemberUsersQueryKey,
-    getAdminServiceListUsergroupMemberUsersQueryKey,
   } from "@rilldata/web-admin/client";
   import DelayedSpinner from "@rilldata/web-common/features/entity-management/DelayedSpinner.svelte";
   import OrgGroupsTable from "@rilldata/web-admin/features/organizations/users/OrgGroupsTable.svelte";
@@ -37,43 +32,16 @@
   const queryClient = useQueryClient();
   const currentUser = createAdminServiceGetCurrentUser();
 
-  const renameUserGroup = createAdminServiceRenameUsergroup();
   const deleteUserGroup = createAdminServiceDeleteUsergroup();
   const addUserGroupRole = createAdminServiceAddOrganizationMemberUsergroup();
   const setUserGroupRole =
     createAdminServiceSetOrganizationMemberUsergroupRole();
   const revokeUserGroupRole =
     createAdminServiceRemoveOrganizationMemberUsergroup();
-  const removeUserGroupMember = createAdminServiceRemoveUsergroupMemberUser();
-  const addUsergroupMemberUser = createAdminServiceAddUsergroupMemberUser();
 
   $: filteredGroups = $listOrganizationMemberUsergroups.data?.members.filter(
     (group) => group.groupName.toLowerCase().includes(searchText.toLowerCase()),
   );
-
-  async function handleRename(groupName: string, newName: string) {
-    try {
-      await $renameUserGroup.mutateAsync({
-        organization: organization,
-        usergroup: groupName,
-        data: {
-          name: newName,
-        },
-      });
-
-      await queryClient.invalidateQueries(
-        getAdminServiceListOrganizationMemberUsergroupsQueryKey(organization),
-      );
-
-      eventBus.emit("notification", { message: "User group renamed" });
-    } catch (error) {
-      console.error("Error renaming user group", error);
-      eventBus.emit("notification", {
-        message: "Error renaming user group",
-        type: "error",
-      });
-    }
-  }
 
   async function handleDelete(deletedUserGroupName: string) {
     try {
@@ -164,68 +132,6 @@
       });
     }
   }
-
-  async function handleRemoveUser(groupName: string, email: string) {
-    try {
-      await $removeUserGroupMember.mutateAsync({
-        organization: organization,
-        usergroup: groupName,
-        email: email,
-      });
-
-      await queryClient.invalidateQueries(
-        getAdminServiceListUsergroupMemberUsersQueryKey(
-          organization,
-          groupName,
-        ),
-      );
-
-      eventBus.emit("notification", {
-        message: "User removed from user group",
-      });
-    } catch (error) {
-      console.error("Error removing user from user group", error);
-      eventBus.emit("notification", {
-        message: "Error removing user from user group",
-        type: "error",
-      });
-    }
-  }
-
-  async function handleAddUsergroupMemberUser(
-    email: string,
-    usergroup: string,
-  ) {
-    try {
-      await $addUsergroupMemberUser.mutateAsync({
-        organization: organization,
-        usergroup: usergroup,
-        email: email,
-        data: {},
-      });
-
-      await queryClient.invalidateQueries(
-        getAdminServiceListOrganizationMemberUsersQueryKey(organization),
-      );
-
-      await queryClient.invalidateQueries(
-        getAdminServiceListUsergroupMemberUsersQueryKey(
-          organization,
-          usergroup,
-        ),
-      );
-
-      eventBus.emit("notification", {
-        message: "User added to user group",
-      });
-    } catch (error) {
-      console.error("Error adding user to user group", error);
-      eventBus.emit("notification", {
-        message: "Error adding user to user group",
-        type: "error",
-      });
-    }
-  }
 </script>
 
 <div class="flex flex-col w-full">
@@ -261,13 +167,10 @@
         data={filteredGroups}
         currentUserEmail={$currentUser.data?.user.email}
         searchUsersList={$listOrganizationMemberUsers.data?.members ?? []}
-        onRename={handleRename}
         onDelete={handleDelete}
         onAddRole={handleAddRole}
         onSetRole={handleSetRole}
         onRevokeRole={handleRevokeRole}
-        onRemoveUser={handleRemoveUser}
-        onAddUser={handleAddUsergroupMemberUser}
       />
     </div>
   {/if}
@@ -276,4 +179,4 @@
 <CreateUserGroupDialog
   bind:open={isCreateUserGroupDialogOpen}
   groupName={userGroupName}
-/> -->
+/>
